requirements.txt
opencv-python
numpy==1.26.4
pillow
cmake
dlib
face_recognition
gunicorn
flask